# TpSwagger/api/swagger/swagger.yaml
swagger: '2.0'
...
info:
  version: "1.0.0"
  title: TimePuncher
  contact:
    name: Hans Tschan
    url: http://timepuncher.ch
    email: hans.tschan@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: localhost:3000
basePath: /api/v1
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json

################################################################################
#                                  Security                                    #
################################################################################



################################################################################
#                                           Paths                              #
################################################################################
paths:
...
  /register:
    post:
      description:
        Ein Benutzer registriert sich mit einem Satz von Registrierungsparametern.
        Bei erfolgreicher Registrierung empfängt der Benutzer eine E-Mail mit
        einem Registrierungslink. Dieser kann einmalig verwendet werden. Danach
        kann sich der Benutzer anmelden.
      summary: Einen Benutzer registrieren [AllowAnonymous]
      operationId: registerUser
      parameters:
        - name: registerVm
          in: body
          required: true
          description: Registrierungsinformationen
          schema:
            $ref: '#/definitions/RegisterVm'
      tags:
        - Authentication
      responses:
        200:
          description: Die Operation war erfolgreich.
            Der Benutzer erhält eine E-Mail mit einem Bestätigungslink.
          schema:
            $ref: '#/definitions/RegisterResponse'
        default:
          description: Unerwarteter Fehler
          schema:
            $ref: '#/definitions/OpResult'
  /confirm:
    post:
      description:
        Diese Methode wird durch den Bestätigungslink aufgerufen.
      summary: Eine Benutzerregistrierung bestätigen [AllowAnonymous]
      operationId: confirmRegister
      parameters:
        - name: confirmToken
          in: query
          type: string
          required: true
          description: Registrierungsbestätigung
      tags:
        - Authentication
      responses:
        200:
          description: Die Operation war erfolgreich.
          schema:
            $ref: '#/definitions/ConfirmResponse'
        default:
          description: Unerwarteter Fehler
          schema:
            $ref: '#/definitions/OpResult'
...
definitions:
...
  RegisterVm:
    type: object
    properties:
      firstname:
        type: string
        description: Der Vorname, 1 .. 80 Zeichen
      name:
        type: string
        description: Der Name, 1 .. 80 Zeichen
      email:
        type: string
        format: email
        description: Die E-Mail Adresse 1 .. 160 Zeichen. Wird benötigt für die Bestätigung der Kontoerstellung.
      username:
        type: string
        description: Der Benutzername 1 .. 80 Zeichen. Wird für das Anmeldung benötigt.
      password:
        type: string
        description: Das Passwort 1 .. 80 Zeichen. Wird für die Anmeldung benötigt.
  RegisterResponse:
    type: object
    properties:
      status:
        $ref: '#/definitions/OpResult'
  ConfirmResponse:
    type: object
    properties:
      status:
        $ref: '#/definitions/OpResult'
...
  OpResult:
    type: object
    properties:
      status:
        type: integer
        description: The http-Code
      success:
        type: boolean
        description: True wenn die Operation erfolgreich war
      result:
        type: object
        description: Eine Text-Meldung